+++ src-bin/opam.ml
@@ -6,57 +6,59 @@
 
 open Bos_setup
 
-let get_pkg_dir p opam_pkg_dir = match opam_pkg_dir with
-| Some d -> Ok d
+let get_opam_publish_file p opam_publish_file = match opam_publish_file with
+| Some file -> Ok file
 | None ->
     Topkg_care.Pkg.build_dir p
     >>= fun bdir -> Topkg_care.Pkg.distrib_filename ~opam:true p
-    >>= fun fname -> Ok Fpath.(bdir // fname)
+    >>= fun fname -> Ok Fpath.(bdir // fname + ".opam")
 
 let descr pkg =
-  Topkg_care.Pkg.opam_descr pkg >>= fun descr ->
+  Topkg_care.Pkg.opam_descr pkg >>= fun (descr, _) ->
   Logs.app (fun m -> m "%s" (Topkg_care.Opam.Descr.to_string descr));
   Ok 0
 
-let pkg pkg dist_pkg opam_pkg_dir =
-  let log_pkg dir =
-    Logs.app (fun m -> m "Wrote opam package %a" Topkg_care.Pp.path dir)
+let pkg pkg dist_pkg opam_publish_file =
+  let log_pkg dst =
+    Logs.app (fun m -> m "Wrote opam package %a" Topkg_care.Pp.path dst)
   in
   let warn_if_vcs_dirty () =
     Cli.warn_if_vcs_dirty "The opam package may be inconsistent with the \
     distribution."
   in
-  get_pkg_dir pkg opam_pkg_dir
-  >>= fun dir -> Topkg_care.Pkg.opam pkg
+  let opam_file_content opam (descr, from_opam) url =
+    let descr =
+      if from_opam then "" else
+      (Topkg_care.Opam.Descr.to_opam_fields descr ^ "\n")
+    in
+    strf "%s\n%s%s" opam descr (Topkg_care.Opam.Url.to_opam_section url)
+  in
+  get_opam_publish_file pkg opam_publish_file
+  >>= fun dst -> Topkg_care.Pkg.opam pkg
   >>= fun opam -> OS.File.read opam
   >>= fun opam -> Topkg_care.Pkg.opam_descr pkg
-  >>= fun descr -> Ok (Topkg_care.Opam.Descr.to_string descr)
   >>= fun descr -> Topkg_care.Pkg.distrib_file dist_pkg
   >>= fun distrib_file -> Topkg_care.Pkg.distrib_uri dist_pkg
   >>= fun uri -> Topkg_care.Opam.Url.with_distrib_file ~uri distrib_file
-  >>= fun url -> OS.Dir.exists dir
-  >>= fun exists -> (if exists then OS.Dir.delete ~recurse:true dir else Ok ())
-  >>= fun () -> OS.Dir.create dir
-  >>= fun _ -> OS.File.write Fpath.(dir / "descr") descr
-  >>= fun () -> OS.File.write Fpath.(dir / "opam") opam
-  >>= fun () -> OS.File.write Fpath.(dir / "url") url
-  >>= fun () -> log_pkg dir; warn_if_vcs_dirty ()
+  >>= fun url -> OS.File.write dst (opam_file_content opam descr url)
+  >>= fun () -> log_pkg dst; warn_if_vcs_dirty ()
   >>= fun () ->
   Ok 0
 
-let submit pkg opam_pkg_dir =
+let submit pkg opam_pkg_dst =
   Topkg_care.Opam.ensure_publish ()
-  >>= fun () -> get_pkg_dir pkg opam_pkg_dir
-  >>= fun pkg_dir -> OS.Dir.exists pkg_dir
+  >>= fun () -> get_opam_publish_file pkg opam_pkg_dst
+  >>= fun opam_file -> OS.File.exists opam_file
   >>= function
   | false ->
-      Logs.err (fun m -> m "Package@ %a@ does@ not@ exist. Did@ you@ forget@ \
-                            to@ invoke 'topkg opam pkg' ?" Fpath.pp pkg_dir);
+      Logs.err (fun m -> m "Package@ file %a@ does@ not@ exist. Did@ you@ \
+                            forget@ to@ invoke 'topkg opam pkg' ?"
+                   Fpath.pp opam_file);
       Ok 1
   | true ->
-      Logs.app (fun m -> m "Submitting %a" Topkg_care.Pp.path pkg_dir);
+      Logs.app (fun m -> m "Publishing %a" Topkg_care.Pp.path opam_file);
       Topkg_care.Pkg.publish_msg pkg
-      >>= fun msg -> Topkg_care.Opam.submit ~pkg_dir ~msg >>= fun () -> Ok 0
+      >>= fun msg -> Topkg_care.Opam.submit ~opam_file ~msg >>= fun () -> Ok 0
 
 let field pkg field = match field with
 | None -> Logs.err (fun m -> m "Missing FIELD positional argument"); Ok 1
@@ -73,7 +75,7 @@ let field pkg field = match field with
 
 let opam () pkg_file build_dir
     dist_name dist_version dist_opam dist_uri dist_file
-    pkg_opam_dir pkg_name pkg_version pkg_opam pkg_descr
+    opam_publish_file pkg_name pkg_version pkg_opam pkg_descr
     readme change_log publish_msg action field_name
   =
   let p =
@@ -90,8 +92,8 @@ let opam () pkg_file build_dir
           ?distrib_uri:dist_uri ?distrib_file:dist_file ?readme ?change_log
           ?publish_msg pkg_file
       in
-      pkg p dist_p pkg_opam_dir
-  | `Submit -> submit p pkg_opam_dir
+      pkg p dist_p opam_publish_file
+  | `Submit -> submit p opam_publish_file
   | `Field -> field p field_name
   end
   |> Cli.handle_error
@@ -102,7 +104,7 @@ open Cmdliner
 
 let action =
   let action = [ "descr", `Descr; "pkg", `Pkg; "submit", `Submit;
-                 "field", `Field ]
+                 "publish", `Submit; "field", `Field ]
   in
   let doc = strf "The action to perform. $(docv) must be one of %s."
       (Arg.doc_alts_enum action)
@@ -114,14 +116,15 @@ let field =
   let doc = "the field to output ($(b,field) action)" in
   Arg.(value & pos 1 (some string) None & info [] ~doc ~docv:"FIELD")
 
-let pkg_opam_dir =
-  let doc = "Directory to use to write the opam package. If absent the
-             directory $(i,BUILD_DIR)/$(i,PKG_NAME).$(i,PKG_VERSION) of the
+let opam_publish_file =
+  let doc = "The file to use to publish the opam package. If absent the
+             file $(i,BUILD_DIR)/$(i,PKG_NAME).$(i,PKG_VERSION).opam in the
              build directory is used (see options $(b,--build-dir),
              $(b,--pkg-name) and $(b,--pkg-version))"
   in
-  let docv = "DIR" in
-  Arg.(value & opt (some Cli.path_arg) None & info ["pkg-opam-dir"] ~doc ~docv)
+  let docv = "FILE" in
+  Arg.(value & opt (some Cli.path_arg) None & info ["opam-publish-file"]
+         ~doc ~docv)
 
 let pkg_version =
   let doc = "The version string $(docv) of the opam package. If absent provided
@@ -140,8 +143,11 @@ let pkg_opam =
   Arg.(value & opt (some Cli.path_arg) None & info ["pkg-opam"] ~doc ~docv)
 
 let pkg_descr =
-  let doc = "The opam descr file to use for the opam package. If absent and
-             the opam file name (see $(b,--pkg-opam)) has a `.opam`
+  let doc = "The opam descr file to use for the opam package. If absent
+             and the opam file has synopsis and description fields this
+             is used for the description. It absent and there are no such
+             fields in the opam file and the opam file name
+             (see $(b,--pkg-opam)) has a `.opam`
              extension, uses an existing file with the same path but a `.descr`
              extension. If the opam file name is `opam` uses a `descr`
              file in the same directory. If these files are not found
@@ -177,7 +183,7 @@ let man =
         "create an opam package description for a distribution.
          The action needs a distribution archive to operate, see
          topkg-distrib(1) or the $(b,--dist-file) option.");
-    `I ("$(b,submit)",
+    `I ("$(b,submit) or $(b,publish)",
         "submits a package created with the action $(b,pkg) the OCaml
          opam repository. Requires the $(b,opam-publish) tool to be
          installed.");
@@ -189,7 +195,7 @@ let cmd =
   let t = Term.(pure opam $ Cli.setup $ Cli.pkg_file $ Cli.build_dir $
                 Cli.dist_name $ Cli.dist_version $ Cli.dist_opam $
                 Cli.dist_uri $ Cli.dist_file $
-                pkg_opam_dir $ Cli.pkg_name $ pkg_version $ pkg_opam $
+                opam_publish_file $ Cli.pkg_name $ pkg_version $ pkg_opam $
                 pkg_descr $ Cli.readme $ Cli.change_log $ Cli.publish_msg $
                 action $ field)
   in