+++ src-care/topkg_care.ml
@@ -10,7 +10,6 @@ module Opam = Topkg_care_opam
 module OCamlbuild = Topkg_care_ocamlbuild
 module OCamlfind = Topkg_care_ocamlfind
 module Archive = Topkg_care_archive
-
 module Pkg = Topkg_care_pkg
 module Delegate = Topkg_care_delegate